% Generated by roxygen2 (4.0.1): do not edit by hand
\name{box_stats}
\alias{box_stats}
\title{Calculates boxplot statistics}
\usage{
box_stats(x, whiskers = c(0.25, 0.975))
}
\arguments{
\item{x}{Vector of values from which to calculate statistics}

\item{whiskers}{A two-element vector of the percentiles for the lower and upper whiskers (c(0.25, 0.975))}
}
\value{
A vector containing the 2.5, 25, 50, 75, 97.5th percentile values and the mean
}
\description{
Calculates boxplot statistics
}

