% Generated by roxygen2 (4.0.1): do not edit by hand
\name{boxplot_y}
\alias{boxplot_y}
\title{Makes horizontal boxplots}
\usage{
boxplot_y(x, ycoord, inhgt = 20, whiskhgt, bcol, bfill, whiskcol, n, pcex)
}
\arguments{
\item{x}{Vector containing, respectively, the 5, 25, 50, 75, 95th percentile values and the mean}

\item{ycoord}{Y coordinate for plotting boxplot}

\item{inhgt}{Height of the interquartile box, scaled to size of device and number of boxes to be plotted}

\item{whiskhgt}{Height of the 95th percentile whiskers, scaled in the same way as inhgt}

\item{bcol}{Color of lines (median line, whiskers and interquartile outline) and mean point}

\item{bfill}{Fill color for interquartile box}

\item{n}{Number of boxplots to be placed on plot}

\item{pcex}{Size of point indicating the mean}
}
\value{
Boxplot
}
\description{
Makes horizontal boxplots
}
\details{
To be used with dummy call to plot(). Specify the number of boxes to be placed on the Y axis
}

