% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cite.soft.R
\name{cite.soft}
\alias{cite.soft}
\title{To create BibTex file for references of software}
\usage{
cite.soft(soft, key, output = "software.bib")
}
\arguments{
\item{soft}{list of citation or package want to reference}

\item{key}{key for inline citation}

\item{output}{output file}
}
\value{
Bibtex file
}
\description{
To create BibTex file for references of software
}

