% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{setdiff, but in both directions}
\usage{
compare(x, y)
}
\arguments{
\item{x}{A vector}

\item{y}{A vector}
}
\value{
A list of:
elements in x but not y
elements in y but not x
union: any element in either x or y
intersection: elements in both x and y
setequal: do both x and y contain the same elements (ignoring duplicates)?
length of x and y
number of unique values in x and in y
}
\description{
setdiff, but in both directions
}
\examples{
compare(1:5, c(3,3,4,4,5,5,6,7))
}
