% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FacialBurns}
\alias{FacialBurns}
\title{Facial Burns Data}
\format{
A data frame with 6 columns and 77 rows
}
\source{
Only for demonstration purposes
}
\usage{
FacialBurns
}
\description{
This dataset was taken from Lavaan 0.6.2. The versions >0.7 have removed it.
}
\keyword{datasets}
