% Generated by roxygen2 (4.0.1): do not edit by hand
\name{flexLegend}
\alias{flexLegend}
\alias{flex_legend}
\title{Creates a vertical or horizontal legend in a location of your choosing}
\usage{
flex_legend(ncuts, legend.text, legend.vals, legend.pos, longdims, shortdims,
  colvec, leg.adj = c(0, 0), cex.val = 1, srt.val = 0, horiz = TRUE,
  textside = "bottom", textcol = "black", bordercol = "black")
}
\arguments{
\item{ncuts}{Number of intervals for legend}

\item{legend.text}{Label for the legend (e.g. units)}

\item{legend.vals}{Vector of labels for the legend}

\item{legend.pos}{A 1 or 2 element vector that specifies where along vector legend.text should be (details)}

\item{longdims}{Vector specifying the user coordinates for the legend's long axis (e.g. c(0.2, 0.8))}

\item{shortdims}{Vector specifying the user coordinates for the legend's short axis (e.g. c(0.01, 0.04))}

\item{colvec}{Vector of colors to fill each polygon in vector, equal in length to ncuts}

\item{leg.adj}{Tweak position(s) of legend.vals and/or legend.text (see details)}

\item{cex.val}{Adjusts the text size of legend.vals and/or legend.text (see details)}

\item{srt.val}{Rotation of legend.vals and/or legend.text (see details)}

\item{horiz}{TRUE (default) or FALSE for horizontal or vertical legend}

\item{textside}{Specify "bottom" (default) or "top" for side of legend.vals on legend}

\item{textcol}{Vector of colors for legend.vals and legend.text}

\item{bordercol}{Outline color for polygon borders in legend}
}
\value{
A polygon based legend
}
\description{
Creates a vertical or horizontal legend in a location of your choosing
}
\details{
This function depends on the TeachingDemos library, and still uses the deprecated cnvrt.coords().
legend.pos uses a two parameter vector, with the second value specifying the polygon number in the legend
next to/above/below which the legend.text will appear, and the first which of the four corners of the
polygon it will be closest to. Default values are chosen if just one (which then specifies polygon number)
or no values are provided. leg.adj is passed as either a list of two two element vectors, or a single two
element vector. If a list, the first element specifies adjustments for legend.values, and the second
element adjustments for the legend.ext. If just a vector, then these values are used for both. e.g.
list(c(0, 0), c(-1, -0.5)); c(0, -0.5).
}

