% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/various.R
\name{full_path}
\alias{full_path}
\title{Easier joining of file paths and file names for reading in files}
\usage{
full_path(path, fname)
}
\arguments{
\item{path}{Path directory}

\item{fname}{File name}
}
\value{
Complete file string
}
\description{
Easier joining of file paths and file names for reading in files
}
\details{
Currently not set up for Windows path variables. Will need an extra line or two.
}

