% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_images_w_path.R
\name{get_images_w_path}
\alias{get_images_w_path}
\title{Create a string vector of all .png files in a folder, including the path}
\usage{
get_images_w_path(path)
}
\arguments{
\item{path}{A path to a folder containing png files.}
}
\value{
A string vector of all .png files in a folder, including the path
}
\description{
Create a string vector of all .png files in a folder, including the path
}
\examples{
\dontrun{
get_images_w_path("path/to/image_folder")
}
}
