% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hba1c_to_percent.R
\name{hba1c_to_percent}
\alias{hba1c_to_percent}
\title{Converts HbA1c from mmol per mol to perecentage}
\usage{
hba1c_to_percent(mmol)
}
\arguments{
\item{percent}{HbA1c value to convert}
}
\value{
HbA1c in percentage
}
\description{
Converts HbA1c from mmol per mol to perecentage
}
\examples{
hba1c_to_percent(mmol=47)
}
