% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hete.plot (Lam’s iMac's conflicted copy 2015-09-17).R, R/hete.plot.R
\name{hete.plot}
\alias{hete.plot}
\title{To produce heterogeneity plot}
\usage{
hete.plot(formula, data, value = NULL)

hete.plot(formula, data, value = NULL)
}
\arguments{
\item{formula}{outcome ~ independent variable}

\item{data}{required data}

\item{value}{value of independent variable}

\item{formula}{outcome ~ independent variable}

\item{data}{required data}

\item{value}{value of independent variable}
}
\value{
graph

graph
}
\description{
To produce heterogeneity plot

To produce heterogeneity plot
}

