% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lmisc-package.R
\docType{package}
\name{Lmisc-package}
\alias{Lmisc}
\alias{Lmisc-package}
\title{Lmisc: Miscellaneous Functions Martin Law}
\description{
Miscellaneous functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/martinlaw/Lmisc}
  \item Report bugs at \url{https://github.com/martinlaw/Lmisc/issues}
}

}
\author{
\strong{Maintainer}: Martin Law \email{martin.law@mrc-bsu.cam.ac.uk}

Authors:
\itemize{
  \item Martin Law \email{martin.law@mrc-bsu.cam.ac.uk}
}

}
\keyword{internal}
