% Generated by roxygen2 (4.0.1): do not edit by hand
\name{loadPath}
\alias{loadPath}
\title{Easier joining of file paths and file names for reading in files}
\usage{
loadPath(path, fname)
}
\arguments{
\item{path}{Path directory}

\item{fname}{File name}
}
\value{
Complete file string
}
\description{
Easier joining of file paths and file names for reading in files
}
\details{
Currently not set up for Windows path variables. Will need an extra line or two.
}

