% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logit}
\alias{logit}
\alias{LO}
\title{Takes the logit. Alias is LO ("log odds").}
\usage{
logit(p)

LO(p)
}
\arguments{
\item{p}{Value for taking logit}
}
\value{
Logit
}
\description{
Takes the logit. Alias is LO ("log odds").
}
\examples{
logit(0.5)
LO(0.5)
}
