% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_performance_criterion.R
\name{myaccuracy}
\alias{myaccuracy}
\title{Functions to assess model performance}
\usage{
myaccuracy(predicted, y, criteria = mse, ...)
}
\arguments{
\item{predicted}{Predicted values from fitted models}

\item{y}{Response}
}
\value{
numeric value of model's performance
}
\description{
Functions to assess model performance
}

