% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mySummary.fit.R
\name{mySummary.fit}
\alias{mySummary.fit}
\alias{mySummary.uni}
\title{To summarize fitted model}
\usage{
mySummary.fit(fit, meta = NULL, exp = FALSE, method = "LRT", digits = 2,
  sep = "-", p = 0.001, tidy = TRUE)

mySummary.uni(fit, adjust = NULL, ...)
}
\arguments{
\item{fit}{fitted models}

\item{meta}{metadata}

\item{exp}{exponential regression coefficients and their CIs?}

\item{method}{estimation method?}

\item{digits}{decimal digit number}

\item{sep}{Separation between lower and upper confidence interval values}

\item{p}{cut-off of p value}

\item{tidy}{Tidy up result?}
}
\value{
Matrix
}
\description{
To summarize fitted model
}
\section{Functions}{
\itemize{
\item \code{mySummary.uni}: To summarize fitted model (univariate analysis)
}}

