% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/various.R
\name{named_out}
\alias{named_out}
\title{Names and returns object inside nested apply functions}
\usage{
named_out(ret, nm)
}
\arguments{
\item{ret}{Object}

\item{nm}{Name to assign to object}
}
\value{
Named object out of loop
}
\description{
This little function assigns a name to an object created by an inner apply function and returns 
it
}
\details{
The entire purpose of this function is simply to save a line when writing nested apply functions 
so that names can be assigned to object and returned without taking up needless space
}
\examples{
ab <- list("a" = c(1:10), "b" = c(11:20))
d <- sample(-5:5, size = 10, replace = TRUE)
lapply(ab$a, function(x) {
  o <- lapply(d, function(j) {
    x * ab$b * d
  })  
  named_out(o, paste("a", x, "b", gsub("-", "neg", d), sep = "")) 
})
}

