% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pchShow}
\alias{pchShow}
\title{Shows the various plotting symbols}
\usage{
pchShow(extras = c("*", ".", "o", "O", "0", "+", "-", "|", "\%", "#"),
  cex = 3, col = "red3", bg = "gold", coltext = "brown",
  cextext = 1.2, main = paste("plot symbols :  points (...  pch = *, cex =",
  cex, ")"))
}
\description{
Shows the various plotting symbols
}
\details{
This was found and copied from here: http://rgraphics.limnology.wisc.edu/pch.php
}

