% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{plot_dag}
\alias{plot_dag}
\title{Plots the dag based on the nodes and edges passed.}
\usage{
plot_dag(
  nodes,
  edges,
  label_size = 16.9,
  label_size_unit = "pt",
  label_border_size = 1,
  text_size = 7,
  xlim = c(0, 14),
  ylim = c(0, 8),
  plot_width_cm = 20,
  plot_height_cm = 12,
  footnote = NULL,
  footnote_size = 12
)
}
\arguments{
\item{nodes}{A dataframe containing data from tidy_dagitty(): node_id, x, y, label}

\item{edges}{A dataframe containing data from tidy_dagitty()}

\item{label_size}{Set the label size of the text inside the nodes}

\item{label_size_unit}{Set the unit of the `label_size`}

\item{label_border_size}{Sets the boxes' border size}

\item{text_size}{Controls the text's size on top of the paths}

\item{xlim}{A vector of integers. Controls the limits of the plot}

\item{ylim}{A vector of integers. Controls the limits of the plot}

\item{plot_width_cm}{The width of the plot in cm}

\item{plot_height_cm}{The height of the plot in cm}

\item{footnote}{Sets the footnote}

\item{footnote_size}{Controls the size of the footnote text}
}
\description{
Plots the dag based on the nodes and edges passed.
}
