% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_ci.R
\name{point_ci}
\alias{point_ci}
\title{Creates typical form for reporting estimate and confidence interval.}
\usage{
point_ci(est, ci.lo, ci.hi, ci.level = "95", ...)
}
\arguments{
\item{est}{Estimate}

\item{ci.lo}{Lower confidence interval}

\item{ci.hi}{Upper confidence interval}

\item{ci.level}{Level of confidence interval (default "95")}

\item{...}{Number of decimal places (passed to function dp())}
}
\value{
Typical character data for estimate and confidence interval.
}
\description{
Creates typical form for reporting estimate and confidence interval.
}
\examples{
point_ci(est=3.001, ci.lo=1.999, ci.hi=5.10001, ci.level="90")
}
