% Generated by roxygen2 (4.0.1): do not edit by hand
\name{proj_root}
\alias{proj_root}
\title{Creates root directory for project (one above project directory), to facilitate projects across computers}
\usage{
proj_root(proj.dir)
}
\arguments{
\item{proj.dir}{Character vector specifying name of project working directory (not a full file path)}
}
\value{
A file path with the platform-specific path structure
}
\description{
Creates root directory for project (one above project directory), to facilitate projects across computers
}

