% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lines2.R
\name{read_lines2}
\alias{read_lines2}
\title{Faster version of readLines}
\usage{
read_lines2(fname)
}
\arguments{
\item{fname}{File name to read}
}
\value{
Vector of strings from input file
}
\description{
Faster version of readLines
}
\author{
mlt
}
\references{
http://www.r-bloggers.com/faster-files-in-r/
}

