% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rect_coords}
\alias{rect_coords}
\title{Finds graphics device coordinates for rectangles that can be used to construct image legend}
\usage{
rect_coords(minWorS, maxEorN, ncuts, EWorNS = "EW", constEWorNS, resEWorNS)
}
\arguments{
\item{minWorS}{The minimum X (horizontal legend) or Y (vertical) user coordinate (0 - 1) for the legend}

\item{maxEorN}{The maximum X (horizontal legend) or Y (vertical) user coordinate (0 - 1) for the legend}

\item{ncuts}{Number of intervals for legend}

\item{EWorNS}{"EW" for horizontal legend (default) or "NS" for vertical}

\item{constEWnorS}{The coordinate that will remain constant (X coord for vertical legend, Y for horiz)}

\item{resEWorNS}{The width in user coordinate for the legend}
}
\value{
List of coordinates for the four corners of each segment of legend, in graphical device coords
}
\description{
Finds graphics device coordinates for rectangles that can be used to construct image legend
}
\details{
This function depends on the TeachingDemos library, and still uses the deprecated cnvrt.coords().
}

