% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/various.R
\name{set_base_path}
\alias{set_base_path}
\title{Gives model base path}
\usage{
set_base_path(currdir = ".")
}
\arguments{
\item{rootdir}{Name of directory you want to be in}
}
\value{
Full path of main project directory
}
\description{
Finds the full base path for the directory you are working in, 
useful for knitting Rmd files.
}

