% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_talk_folder.R
\name{set_talk_folder}
\alias{set_talk_folder}
\title{Set path to subfolder within "other people's talks" folder}
\usage{
set_talk_folder(subfolder)
}
\arguments{
\item{subfolder}{A subfolder inside the "other people's talks" folder}
}
\value{
The path to the "other people's talks" folder with the subfolder appended.
}
\description{
Set path to subfolder within "other people's talks" folder
}
\examples{
\dontrun{
set_talk_folder("path/to/folder")
}
}
