% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/timeplot.select.R
\name{timeplot.select}
\alias{timeplot.select}
\title{To systematically select id in a longitudinal dataset based on residuals from fitted smooth}
\usage{
timeplot.select(formula, id, data, smoother = c("loess", "spline"),
  select = c(0, 0.05, 0.1, 0.25, 0.5, 0.75, 0.9, 0.95, 1), df = 4, ...)
}
\arguments{
\item{formula}{formula to supply to smoother (y ~ x)}

\item{id}{character specify name of id variable}

\item{data}{longitudinal data}

\item{smoother}{type of smooth}

\item{select}{quantile to choose}

\item{df}{degree of freedom if use spline as smoother (default = 4)}

\item{...}{other arguments to supply to loess}
}
\value{
data frame contains id, max x, y at max x, and labels
}
\description{
To systematically select id in a longitudinal dataset based on residuals from fitted smooth
}

