% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/welch.R
\name{welch}
\alias{welch}
\title{Undertake Welch test using means, standard deviations and sample sizes}
\usage{
welch(mu1, mu2, sd1, sd2, n1, n2, two.sided.alpha = 0.05)
}
\arguments{
\item{mu1}{Mean of group 1}

\item{mu2}{Mean of group 2}

\item{sd1}{SD of group 1}

\item{sd2}{SD of group 2}

\item{n1}{Sample size of group 1}

\item{n2}{Sample size of group 2}

\item{two.sided.alpha}{Two-sided alpha (default 0.05)}
}
\value{
Vector of t-statistic, estimate, confidence interval and p-value
}
\description{
Undertake Welch test using means, standard deviations and sample sizes
}
\examples{
welch(mu1=mean(1:10), mu2=mean(7:20), sd1=sd(1:10), sd2=sd(7:20), n1=length(1:10), n2=length(7:20))
}
