% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp.R
\name{dp}
\alias{dp}
\title{Return number with a certain number of decimal places}
\usage{
dp(x, digits = 2)
}
\arguments{
\item{x}{A number}

\item{digits}{Number of decimal places to give (default 2)}
}
\value{
Character: the number to the required number of decimal places.
}
\description{
Return number with a certain number of decimal places
}
\examples{
dp(1.002)
}
