% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ci.R
\name{find_ci}
\alias{find_ci}
\title{Obtain simple confidence interval}
\usage{
find_ci(est, se, crit = qnorm(0.975), exp = FALSE)
}
\arguments{
\item{est}{Estimate}

\item{se}{Standard error}

\item{crit}{Critical value. Default qnorm(0.975)}

\item{exp}{Logical. Exponentiate the output? Default FALSE}
}
\value{
Single-row data frame containing estimate and confidence intervals.
}
\description{
Obtain simple confidence interval
}
\examples{
find_ci(est=1, se=0.1, crit=qnorm(0.95))
}
