% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ci.R
\name{find_ci}
\alias{find_ci}
\title{Obtain simple confidence interval}
\usage{
find_ci(
  est,
  lo.ci = NULL,
  hi.ci = NULL,
  se = NULL,
  crit = stats::qnorm(0.975),
  exp = FALSE,
  recip = FALSE,
  digits = NULL
)
}
\arguments{
\item{est}{Estimate}

\item{lo.ci}{Lower CI}

\item{hi.ci}{Upper CI}

\item{se}{Standard error}

\item{crit}{Critical value. Default qnorm(0.975)}

\item{exp}{Logical. Exponentiate the output? Default FALSE}

\item{recip}{Take reciprocal}

\item{digits}{Number of decimal places}
}
\value{
Single-row data frame containing estimate and confidence intervals.
}
\description{
Obtain simple confidence interval
}
\examples{
find_ci(est=1, se=0.1, crit=1.96)
}
