% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hba1c_to_mmol.R
\name{hba1c_to_mmol}
\alias{hba1c_to_mmol}
\title{Converts HbA1c from percentage to mmol per mol}
\usage{
hba1c_to_mmol(percent)
}
\arguments{
\item{percent}{HbA1c value to convert}
}
\value{
HbA1c in mmol per mol
}
\description{
Converts HbA1c from percentage to mmol per mol
}
\examples{
hba1c_to_mmol(percent=7)
}
