\name{Lmisc-package}
\alias{Lmisc-package}
\alias{Lmisc}
\docType{package}
\title{
Collection of my miscellaneous functions
}
\description{
Collection of my miscellaneous functions
}
\details{
\tabular{ll}{
Package: \tab Lmisc\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-07-23\cr
License: \tab GPL-2\cr
}
At the moment there is only 2 functions: createProject, get.dropbox.folder
}
\author{
Phung Khanh Lam
Maintainer: Phung Khanh Lam <phungkhanhlam@gmail.com>
}
\references{
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
get.dropbox.folder()
}
