% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_ci.R
\name{print_ci}
\alias{print_ci}
\title{Prints typical form for reporting estimate and confidence interval, with optional rounding.}
\usage{
print_ci(est, lo.ci, hi.ci, ci.level = "95", digits = 2)
}
\arguments{
\item{est}{Estimate}

\item{lo.ci}{Lower confidence interval}

\item{hi.ci}{Upper confidence interval}

\item{ci.level}{Level of confidence interval (default "95")}

\item{digits}{Number of decimal places (default NULL)}
}
\value{
Typical character data for estimate and confidence interval.
}
\description{
Prints typical form for reporting estimate and confidence interval, with optional rounding.
}
\examples{
print_ci(est=3.001, lo.ci=1.999, hi.ci=5.10001, ci.level="90", digits=2)
}
