% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_p.R
\name{print_p}
\alias{print_p}
\title{Print p value as-is, or with "<0.x", to selected decimal place}
\usage{
print_p(p, decimals = 2, sep = "", add.p = FALSE)
}
\arguments{
\item{p}{P-value to print}

\item{decimals}{Number of decimal places (default=3)}

\item{sep}{Separator on each side of operator (default="")}

\item{add.p}{Logical. Should "p" or "p=" be included in outout? Default FALSE.}
}
\value{
String showing p-value
}
\description{
Print p value as-is, or with "<0.x", to selected decimal place
}
\examples{
print_p(0.000004321, dec=4)
}
