% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_path.R
\name{set_path}
\alias{set_path}
\title{Create character for path to project}
\usage{
set_path(project.number)
}
\arguments{
\item{project.number}{Project number}
}
\value{
Character for path to project.
}
\description{
Create character for path to project
}
\examples{
set_path(135)
}
